ARM: dts: sun5i: hsg-h702: add axp209 regulator nodes
This patch adds the regulator nodes for the axp209. As the
inputs of these regulators are from the axp209's PS output,
which is basically just a mux over the 2 inputs, it is
considered to be unregulated. Thus we do not provide input
supply properties for them.

The regulator names and constraints are based on the reference
design schematics and the SoC datasheet, with testing to confirm.

DCDC2 is used as the cpu power supply. This patch also references
it from the cpu node.

LDO3 powers the USB WiFi module. This patch also references
it from the usb-phy node.

Signed-off-by: Chen-Yu Tsai <wens@csie.org>