ARM: dts: sun8i: Add cpu clock reference and operating points to dtsi
The cpu core is clocked from the "cpu" clock. Add a reference to it
in the first cpu node. Also add "cpu0" label to the node.

The operating points were taken from the A23 FEX files in the
sunxi-boards repository. Currently all versions of Ippo-q8h have
the same settings.

Signed-off-by: Chen-Yu Tsai <wens@csie.org>