ARM: dts: sunxi: Add dtsi file for AXP806 PMIC
The AXP806 PMIC is used with the Allwinner A80 SoC, along with
an AXP809 PMIC as a master.

This patch adds a dtsi file for all the common bindings and default
values unrelated to board design. Currently this is just listing all
the regulator nodes. The regulators are initialized based on their
device node names.

Signed-off-by: Chen-Yu Tsai <wens@csie.org>